Changed int to int16_t for 32 bit Arduino
Not yet verified on Due

+++ ADXL362.cpp
@@ -21,7 +21,7 @@
 
 //#define ADXL362_DEBUG
 
-int slaveSelectPin = 10;
+int16_t slaveSelectPin = 10;
 
 ADXL362::ADXL362() {
 }
@@ -31,7 +31,7 @@ ADXL362::ADXL362() {
 //  begin()
 //  Initial SPI setup, soft reset of device
 //
-void ADXL362::begin(int chipSelectPin) {
+void ADXL362::begin(int16_t chipSelectPin) {
 	slaveSelectPin = chipSelectPin;
 	pinMode(slaveSelectPin, OUTPUT);
 	SPI.begin();
@@ -75,8 +75,8 @@ void ADXL362::beginMeasure() {
 //  readXData(), readYData(), readZData(), readTemp()
 //  Read X, Y, Z, and Temp registers
 //
-int ADXL362::readXData(){
-	int XDATA = SPIreadTwoRegisters(0x0E);
+int16_t ADXL362::readXData(){
+	int16_t XDATA = SPIreadTwoRegisters(0x0E);
 	
 #ifdef ADXL362_DEBUG
 	Serial.print("XDATA = ");
@@ -86,8 +86,8 @@ int ADXL362::readXData(){
 	return XDATA;
 }
 
-int ADXL362::readYData(){
-	int YDATA = SPIreadTwoRegisters(0x10);
+int16_t ADXL362::readYData(){
+	int16_t YDATA = SPIreadTwoRegisters(0x10);
 
 #ifdef ADXL362_DEBUG
 	Serial.print("\tYDATA = "); 
@@ -97,8 +97,8 @@ int ADXL362::readYData(){
 	return YDATA;
 }
 
-int ADXL362::readZData(){
-	int ZDATA = SPIreadTwoRegisters(0x12);
+int16_t ADXL362::readZData(){
+	int16_t ZDATA = SPIreadTwoRegisters(0x12);
 
 #ifdef ADXL362_DEBUG
 	Serial.print("\tZDATA = "); 
@@ -108,8 +108,8 @@ int ADXL362::readZData(){
 	return ZDATA;
 }
 
-int ADXL362::readTemp(){
-	int TEMP = SPIreadTwoRegisters(0x14);
+int16_t ADXL362::readTemp(){
+	int16_t TEMP = SPIreadTwoRegisters(0x14);
 
 #ifdef ADXL362_DEBUG
 	Serial.print("\tTEMP = "); 
@@ -119,7 +119,7 @@ int ADXL362::readTemp(){
 	return TEMP;
 }
 
-void ADXL362::readXYZTData(int &XData, int &YData, int &ZData, int &Temperature){
+void ADXL362::readXYZTData(int16_t &XData, int16_t &YData, int16_t &ZData, int16_t &Temperature){
 	  // burst SPI read
 	  // A burst read of all three axis is required to guarantee all measurements correspond to same sample time
 	  digitalWrite(slaveSelectPin, LOW);
@@ -143,7 +143,7 @@ void ADXL362::readXYZTData(int &XData, int &YData, int &ZData, int &Temperature)
 #endif
 }
 
-void ADXL362::setupDCActivityInterrupt(int threshold, byte time){
+void ADXL362::setupDCActivityInterrupt(int16_t threshold, byte time){
 	//  Setup motion and time thresholds
 	SPIwriteTwoRegisters(0x20, threshold);
 	SPIwriteOneRegister(0x22, time);
@@ -161,7 +161,7 @@ void ADXL362::setupDCActivityInterrupt(int threshold, byte time){
 #endif
 }
 
-void ADXL362::setupACActivityInterrupt(int threshold, byte time){
+void ADXL362::setupACActivityInterrupt(int16_t threshold, byte time){
 	//  Setup motion and time thresholds
 	SPIwriteTwoRegisters(0x20, threshold);
 	SPIwriteOneRegister(0x22, time);
@@ -179,7 +179,7 @@ void ADXL362::setupACActivityInterrupt(int threshold, byte time){
 #endif
 }
 
-void ADXL362::setupDCInactivityInterrupt(int threshold, int time){
+void ADXL362::setupDCInactivityInterrupt(int16_t threshold, int16_t time){
 	// Setup motion and time thresholds
 	SPIwriteTwoRegisters(0x23, threshold);
 	SPIwriteTwoRegisters(0x25, time);
@@ -197,7 +197,7 @@ void ADXL362::setupDCInactivityInterrupt(int threshold, int time){
 #endif
 }
 
-void ADXL362::setupACInactivityInterrupt(int threshold, int time){
+void ADXL362::setupACInactivityInterrupt(int16_t threshold, int16_t time){
 	//  Setup motion and time thresholds
 	SPIwriteTwoRegisters(0x23, threshold);
 	SPIwriteTwoRegisters(0x25, time);
@@ -266,8 +266,8 @@ void ADXL362::SPIwriteOneRegister(byte regAddress, byte regValue){
 	digitalWrite(slaveSelectPin, HIGH);
 }
 
-int ADXL362::SPIreadTwoRegisters(byte regAddress){
-	int twoRegValue = 0;
+int16_t ADXL362::SPIreadTwoRegisters(byte regAddress){
+	int16_t twoRegValue = 0;
   
 	digitalWrite(slaveSelectPin, LOW);
 	SPI.transfer(0x0B);  // read instruction
@@ -279,7 +279,7 @@ int ADXL362::SPIreadTwoRegisters(byte regAddress){
 	return twoRegValue;
 }  
 
-void ADXL362::SPIwriteTwoRegisters(byte regAddress, int twoRegValue){
+void ADXL362::SPIwriteTwoRegisters(byte regAddress, int16_t twoRegValue){
 	byte twoRegValueH = twoRegValue >> 8;
 	byte twoRegValueL = twoRegValue;
